Improve PyTorch playground quality
