+++ src/agilab/apps/builtin/pytorch_playground_project/src/app_settings.toml
@@ -28,7 +28,7 @@ scheduler = "127.0.0.1:8786"
 workers_data_path = ""
 
 [cluster.workers]
-"127.0.0.1" = 2
+"127.0.0.1" = 1
 
 [cluster.service_health]
 allow_idle = false
